[MISC] Rename filter.* to gamestate_filter.*
